Auth & Guest, Relationships, Add Ownership to Listings, Manage Listings Page, User Authorization

+++ app/Http/Controllers/ListingController.php
@@ -39,6 +39,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos','public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
         return redirect('/')->with('message', 'Listing created successfully');
     }
@@ -48,6 +50,11 @@ class ListingController extends Controller
     }
 
     public function update(Request $request, Listing $listing){
+
+        if($listing->user_id != auth()->id()){
+            abort(403,'Unauthorized action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -66,8 +73,17 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing){
+
+        if($listing->user_id != auth()->id()){
+            abort(403,'Unauthorized action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
 
+    }
+
+    public function manage(){
+        return view('listings.manage',['listings'=>auth()->user()->listings()->get()]);
     }
 }
